fix buffer overflow found by fuzzer
The buffer overflow happens much later, but it is caused by some logic trying to play move on a position where there's already a piece, which increments the disks_played variable above 61, which is then used as index to some array. Some plays doesn't invoke the buffer overflow, but only lead to inconsistent state.